+++ src/App.js
@@ -2,21 +2,21 @@ import "./assets/css/reset.css";
 import styled from "styled-components";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { useState } from "react";
-import UserContext from "./contexts/UserContext.js";
+import UserContext from "./../contexts/UserContext.js";
 
 import TestPage from "./components/TestPage.js";
 import Login from "./components/Login";
 import SignUp from "./components/Sign-Up";
 
-function Error(e) {
-    console.log(`${e.response.status} - ${e.response.statusText}`);
-    alert("Um erro aconteceu, tente novamente");
-  }
 
 export default function App() {
 
     const [user, setUser] = useState({});
 
+    function Error(e) {
+        console.log(`${e.response.status} - ${e.response.statusText}`);
+        alert("Um erro aconteceu, tente novamente");
+      }
 
     return (
         <Div>
